Update FirebaseService to use UrlLauncher and modify related pages
This commit updates the FirebaseService class to include a method for opening links using the UrlLauncher package. It also modifies the ForceUpdate and UpdateFound pages to use this new method. Additionally, the necessary changes are made in the generated_plugin_registrant files for different platforms (linux, macos, windows) to include UrlLauncherPlugin.

Changes in the commit include:

- Adding UrlLauncher as a dependency in pubspec.yaml
- Modifying FirebaseService to include a method for opening links using UrlLauncher
- Updating ForceUpdate and UpdateFound pages to use the new FirebaseService method for opening links
- Adding UrlLauncherPlugin registration in generated_plugin_registrant files for linux, macos, and windows platforms